feat: 로또 키오스크 기능 구현

+++ src/main/java/org/example/Lottery.java
@@ -107,7 +107,7 @@ public class Lottery extends Gamble{
         int totalPrize = 0;
         for(int i =0; i<result.length; i++){
             sb.append( String.format("%d개 일치한 개수 : %d 개\n", i+3, result[i] ));
-            totalPrize += result[i] * (5000 * Math.pow(5000,i));
+            totalPrize += result[i] * (5000 * (int)Math.pow(5000,i));
         }
         sb.append(String.format("\n총 상금은 %d원 입니다.", totalPrize));
         return sb;
